+++ src/main.rs
@@ -14,18 +14,18 @@ $ cargo install caminos
 
 Then you should  be able to run it.
 ```bash
-caminos --help
+$ caminos --help
 ```
 
 To perform a single simulation create a file `my_experiment.cfg` and execute the simulator. See the `caminos-lib` crate for documentation on its format.
 ```bash
-caminos my_experiment.cfg
+$ caminos my_experiment.cfg
 ```
 You may set the `--results` flags to write the simulation result into it instead of to stdout.
 
 For more complex experiments it is recommended to make a new directory `/path/to/my/experiment`. This directory should include a file `main.cfg` describing the experiment to perform and a file `main.od` describing the outputs to be generated. It may contain a file `remote` to help to `pull` result files launched remotely. Then, to run all the simulations locally and create the outputs execute the following.
 ```bash
-caminos /path/to/my/experiment
+$ caminos /path/to/my/experiment
 ```
 
 # Executing Simulations Using SLURM
@@ -62,6 +62,14 @@ local$ caminos /path/to/my/experiment --action=pull
 ```bash
 local$ caminos /path/to/my/experiment --action=output
 ```
+
+# Special modes
+
+The `special` flag enables extra modes other than simulating scenarios. Currently the only such availale mode is exporting a topology. By setting `--special=export` and `--special_args='Export{...}'` it will create a topology file. An example command is the following.
+
+```bash
+$ caminos --special=export --special_args='Export{topology:RandomRegularGraph{routers:500,degree:20,servers_per_router:1},seed:5,filename:"the_topology_file"}'
+```
 */
 
 use std::env;
@@ -69,12 +77,77 @@ use std::fs::{File};
 use std::path::{Path};
 use getopts::{Options};
 use std::str::FromStr;
+use std::cell::{RefCell};
+use rand::{StdRng,SeedableRng};
 
 use caminos_lib::{get_git_id,directory_main,file_main,Plugs,
-	//topology::{self,Topology},
-	//config_parser::ConfigurationValue,
+	topology::{self},
+	config_parser::{self,ConfigurationValue},
 	experiments::{Action,ExperimentOptions},
 	};
+fn special_export(args: &str, plugs:&Plugs)
+{
+	let topology_cfg = match config_parser::parse(args)
+	{
+		Ok(x) => match x
+		{
+			config_parser::Token::Value(value) => value,
+			_ => panic!("Not a value"),
+		},
+		Err(x) => panic!("Error parsing topology to export ({:?})",x),
+	};
+	let mut topology = None;
+	let mut seed = None;
+	let mut format = None;
+	let mut filename = None;
+	if let ConfigurationValue::Object(ref cv_name, ref cv_pairs)=topology_cfg
+	{
+		if cv_name!="Export"
+		{
+			panic!("A Export must be created from a `Export` object not `{}`",cv_name);
+		}
+		for &(ref name,ref value) in cv_pairs
+		{
+			//match name.as_ref()
+			match AsRef::<str>::as_ref(&name)
+			{
+				"topology" =>
+				{
+					topology=Some(value);
+				},
+				"seed" => match value
+				{
+					&ConfigurationValue::Number(f) => seed=Some(f as usize),
+					_ => panic!("bad value for seed"),
+				},
+				"format" => match value
+				{
+					&ConfigurationValue::Number(f) => format=Some(f as usize),
+					_ => panic!("bad value for format"),
+				},
+				"filename" => match value
+				{
+					&ConfigurationValue::Literal(ref s) => filename=Some(s.to_string()),
+					_ => panic!("bad value for filename"),
+				},
+				_ => panic!("Nothing to do with field {} in Export",name),
+			}
+		}
+	}
+	else
+	{
+		panic!("Trying to create a Export from a non-Object");
+	}
+	let seed=seed.unwrap_or(42);
+	let topology_cfg=topology.expect("There were no topology.");
+	let format=format.unwrap_or(0);
+	let filename=filename.expect("There were no filename.");
+	let rng=RefCell::new(StdRng::from_seed(&[seed]));
+	let topology = topology::new_topology(topology::TopologyBuilderArgument{cv:&topology_cfg,plugs,rng:&rng});
+	let mut topology_file=File::create(&filename).expect("Could not create topology file");
+	topology.write_adjacencies_to_file(&mut topology_file,format).expect("Failed writing topology to file");
+}
+
 
 fn main()
 {
@@ -95,8 +168,11 @@ fn main()
 	opts.optopt("r","results","file in which to write the simulation results (for file experiment)","FILE");
 	opts.optopt("s","start_index","experiment index in which to start processing","INDEX");
 	opts.optopt("e","end_index","experiment index in which to end processing","INDEX");
-	//opts.optopt("x","special","some special execution","SPECIAL_VALUE");
+	opts.optopt("x","special","some special execution","SPECIAL_VALUE");
+	opts.optopt("","special_args","arguments for special execution","SPECIAL_VALUE");
 	opts.optopt("f","source","copy matching results from another path experiment","PATH");
+	opts.optopt("w","where","select the subset of indices for which the configuration expression evaluates to true","EXPRESION");
+	opts.optopt("m","message","write a message into the journal file","TEXT");
 	opts.optflag("h","help","show this help");
 	let option_matches= match opts.parse(&args[1..])
 	{
@@ -104,28 +180,29 @@ fn main()
 		Err(f) => panic!(f.to_string()),
 	};
 
-	//if option_matches.opt_present("special")
-	//{
-	//	let special_str = option_matches.opt_str("special").expect("no special string");
-	//	match special_str.as_ref()
-	//	{
-	//		"xxx" =>
-	//		{
-	//			return;
-	//		},
-	//		_ => panic!("unrecognized special function {}",special_str),
-	//	}
-	//}
-
-	if option_matches.free.is_empty() || option_matches.opt_present("help")
+	if (option_matches.free.is_empty() && !option_matches.opt_present("special")) || option_matches.opt_present("help")
 	{
 		let brief = format!("Use:\n\t{binary} configuration_filename [-h] [--special=<method>] [--results=<path>]\n\t{binary} experiment_folder [--action=<method>] [--start_index=<index>] [--end_index=<index>] [--source=<path>]",binary=args[0]);
 		print!("{}", opts.usage(&brief));
 		return;
 	}
 
-	let path=Path::new(&option_matches.free[0]);
 	let plugs = Plugs::default();
+	if option_matches.opt_present("special")
+	{
+		let special_str = option_matches.opt_str("special").expect("no special string");
+		match special_str.as_ref()
+		{
+			"export" =>
+			{
+				special_export(&option_matches.opt_str("special_args").unwrap(),&plugs);
+				return;
+			},
+			_ => panic!("unrecognized special function {}",special_str),
+		}
+	}
+
+	let path=Path::new(&option_matches.free[0]);
 	if path.is_dir()
 	{
 		let action=if option_matches.opt_present("action")
@@ -148,6 +225,19 @@ fn main()
 		if option_matches.opt_present("end_index")
 		{
 			options.end_index = Some(option_matches.opt_str("end_index").unwrap().parse::<usize>().expect("non-usize received from --end_index"));
+		}
+		if option_matches.opt_present("where")
+		{
+			let expr = match config_parser::parse_expression(&option_matches.opt_str("where").unwrap()).expect("error parsing the where clause")
+			{
+				config_parser::Token::Expression(expr) => expr,
+				x => panic!("the where clause is not an expression ({:?}), which it should be.",x),
+			};
+			options.where_clause = Some(expr);
+		}
+		if option_matches.opt_present("message")
+		{
+			options.message = Some(option_matches.opt_str("message").unwrap());
 		}
 		return directory_main(&path,&args[0],&plugs,action,options);
 	}