+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Change Log
 
+## [0.2.0]
+
+### 2021-03-12
+Updating toml to actually use caminos-lib-0.2
+git commit -m "Commit before publishing version 0.2.0 to crates.io."
+
+### 2021-02-12
+Added --message=text and --where=expression flags.
+Added special mode to export topologies.
+Updated to use caminos-lib-0.2
+
+
 ## [0.1.0]
 
-# 2020-11-24
+### 2020-11-24
 git commit -m "First commit on the publicable version."